+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/machine/incubator/IncubatorMenu.kt
@@ -73,7 +73,7 @@ class IncubatorMenu(
         return if (progress == 0) {
             0
         } else {
-            progress * progressArrowSize / IncubatorBlockEntity.ticksPerBrew
+            progressArrowSize - progress * progressArrowSize / IncubatorBlockEntity.ticksPerBrew
         }
     }