+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/machine/incubator_advanced/AdvancedIncubatorMenu.kt
@@ -82,7 +82,7 @@ class AdvancedIncubatorMenu(
         return if (progress == 0) {
             0
         } else {
-            progress * progressArrowSize / IncubatorBlockEntity.ticksPerBrew
+            progressArrowSize - progress * progressArrowSize / IncubatorBlockEntity.ticksPerBrew
         }
     }