+++ CHANGELOG.md
@@ -96,5 +96,6 @@
 ### 1.1.8
 
 - Fixed crash from trying to scrape invalid entities
-- Fixed the Coal Generator, so FE goes up instead of down
+- Fixed the Coal Generator's burn progress rendering incorrectly
+- Fixed the progress arrows in the Incubator and Advanced Incubator rendering incorrectly
 - All machine Screen classes now extend the abstract class MachineScreen, vastly reducing code duplication